Commit dei limiti modificati al termine di una transazione. Ricaricamento limiti specifici sul cliente. Controllo e rimozione limiti duplicati. refs #15

+++ src/structs.h
@@ -89,7 +89,7 @@ struct UserInfo
 
 struct Limit
 {
-    qulonglong id;
+    qlonglong id;
     QString clientCode;
     QString clientTag;
     QString productCode;
@@ -97,7 +97,7 @@ struct Limit
     double limit;
     double current;
     int priority;
-    int parent;
+    qlonglong parent;
 };
 
 struct ClientInfo
